Biggest commit so far
This commit includes all the changes I made while trying to get VSCode to correctly debug and run my code to output lines of text.  There is a known issue with C# and VSCode that forced me to switch to Visual Studio, which I will be working with for the foreseeable future.